refactor: Replace BonusBouton component with BonusButton component
